requirements_dev.txt
# Test framework
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-homeassistant-custom-component>=0.13.0

# HTTP mocking for API client tests
aioresponses>=0.7.6

# Linting & formatting
ruff>=0.6.0
